chore: move metamorphosis module
