+++ docs/module-structure.md
@@ -15,5 +15,6 @@ The project organizes game logic under the `game/` directory.
 - **game/tooltip.js** – simple helpers for showing and hiding the UI tooltip.
 - **game/sect.js** – implements the sect management system including constructs and colony helpers.
 - **game/buildings.js** – handles sect building logic and construction timers.
+- **game/metamorphosis.js** – manages disciple metamorphosis progression and its UI.
 
 The top-level `script.js` now merely orchestrates initialization, delegating all logic to the modules above.
