+++ game/sect.js
@@ -1,6 +1,6 @@
 /* global updateSectDisplay */
 import addLog from '../log.js';
-import { refreshMetamorphosis, tickMetamorphosis } from '../metamorphosis.js';
+import { refreshMetamorphosis, tickMetamorphosis } from './metamorphosis.js';
 import { sectState, currentEnemy, FRUIT_GROWTH_RATES } from './state.js';
 import { generateDiscipleAttributes } from '../discipleAttributes.js';
 import Disciple from '../disciple.js';
